ExecGit_MVP/requirements.txt
fastapi
uvicorn
python-multipart
gitpython
openai
langchain
pydantic
beautifulsoup4
requests
